split payload
